requirements.txt
dotenv
peewee
python-telegram-bot
pytest-cov
pytest
coverage
